requirements.txt
python-telegram-bot==21.1.1
amightygirl.paapi5-python-sdk==1.0.0